@@ -4,8 +4,9 @@
   "description": "A simple starter for new vue project.",
   "main": "src/main.js",
   "scripts": {
-    "dev": "gulp",
-    "publish_bundle": "gulp publish_bundle"
+    "cnpm-install": "cnpm install",
+    "dev": "vbuild -dc",
+    "build": "node build && vbuild -c && node build/publish-cdn.js"
   },
   "keywords": [
     "vue",
@@ -17,51 +18,16 @@
   "author": "卓文理",
   "license": "MIT",
   "devDependencies": {
-    "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-stage-2": "^6.13.0",
-    "babel-runtime": "^6.11.6",
-    "browser-sync": "^2.18.5",
-    "css-loader": "^0.25.0",
-    "del": "^2.2.2",
-    "eslint": "^3.7.0",
-    "eslint-config-airbnb": "^12.0.0",
-    "eslint-config-vue": "^1.1.0",
-    "eslint-plugin-vue": "^0.1.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "gulp": "^3.9.1",
-    "gulp-eslint": "^3.0.1",
-    "gulp-load-plugins": "^1.4.0",
-    "gulp-notify": "^2.2.0",
-    "gulp-sequence": "^0.4.6",
-    "gulp-util": "^3.0.8",
-    "gulp-watch": "^4.3.11",
-    "html-loader": "^0.4.4",
-    "html-webpack-plugin": "^2.24.1",
-    "inject-loader": "^2.0.1",
+    "cli-color": "^1.2.0",
     "jade": "^1.11.0",
-    "json-loader": "^0.5.4",
-    "less-loader": "^2.2.3",
-    "node-sass": "^3.8.0",
-    "ora": "^0.4.0",
-    "pug": "^2.0.0-beta6",
-    "raw-loader": "^0.5.1",
-    "request": "^2.79.0",
-    "require-dir": "^0.3.1",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "vue-hot-reload-api": "^2.0.6",
-    "vue-html-loader": "^1.2.3",
-    "vue-loader": "^8.5.4",
-    "vue-style-loader": "^1.0.0",
-    "webpack": "^1.13.2",
-    "yargs": "^6.5.0"
+    "node-sass": "^4.5.2",
+    "pug": "^2.0.0-beta11",
+    "sass-loader": "^6.0.3",
+    "vbuild": "^6.24.14",
+    "vue-template-compiler": "^2.2.6"
   },
   "dependencies": {
+    "babel-preset-es2015": "^6.24.0",
     "bluebird": "^3.4.6",
     "element-ui": "^1.0.9",
     "highlight.js": "^9.7.0",
@@ -70,7 +36,7 @@
     "marked": "^0.3.6",
     "material-design-iconfonts": "^1.0.0",
     "sass-core": "^1.0.0",
-    "vue": "2.1.3",
+    "vue": "^2.2.6",
     "vue-clipboards": "^0.2.5",
     "vue-router": "^2.0.1",
     "vuex": "^2.0.0"
